refactor: move classes to it\'s respective file
