+++ lib/Character/Character.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+import 'CharacterState.dart';
+
+class Character extends StatefulWidget {
+  CharacterState createState() => new CharacterState();
+}
